Moving arg parsing to their own files
